fix(frontend): make Help dialog responsive on narrow viewports
